requirements.txt
fastapi
openapi
uvicorn
pinecone
neo4j
python-dotenv
spacy
#en-core-web-sm
streamlit
requests